requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.3
pydantic==2.5.0
pydub==0.25.1
soundfile==0.12.1
numpy==1.24.3
librosa==0.10.1
# noisereduce 3.x pulls in torch at import time; stick to 2.0.1 which has no heavy deps
noisereduce==2.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.1.4
pytesseract==0.3.10
pdf2image==1.17.0
pypdf==3.17.4
firebase-admin==6.4.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
google-generativeai==0.3.2
elevenlabs==0.2.26
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3
python-jose[cryptography]==3.3.0
fpdf2==2.7.9
